Update OpenCV 2 code to OpenCV 4
Unfortunately, the C interfaces to access the webcam have been removed.
They were deprecated some time ago and now they are empty functions that
print an error message.

In order to still have GB Camera support it is needed to do one of:

 a) Update the OpenCV code to work with version 4

 b) Find another cross platform library to access the webcam.

Option b has proven to be more difficult than expected, so this commit
does a. It implements a C wrapper over the C++ interface of OpenCV 4.

This should work in all platforms supported by OpenCV, just like before.

The interface of the C wrapper could be improved, though, but at least
the code compiles and runs again, even in modern Linux distros that
don't have old versions of OpenCV in their repositories.