Dynamically allocate framebuffers for debugger windows
This prevents allocating huge buffers on the stack, which is what is
being done at the moment.

In the case of the GB Camera viewer, the buffer was a static buffer,
which meant it was always allocated even if the user wasn't using that
viewer. It's better to only allocate it if needed.