Refactor sound generation
Previously, each console core used to sample the sound at 22050 Hz, and the
output was sent directly to SDL. However, this isn't very portable, as the
sample rate returned by SDL may not be 22050.

The new code generates samples at 32768 Hz in both the GB and GBA cores,
which is sent to an SDL AudioStream, which resamples it to whichever sample
rate is chosen by SDL as output sample rate.

This makes it a lot easier to record accurate WAVs from the emulator cores,
because the 32768 Hz audio is deterministic. The step that resamples it to
send it to the output may need to drop samples every now and then, but the
output of the cores is always the same.