Revert "Don't override key input state from Lua script"
This reverts commit d8a53ce711074147e9fd0664e2830c5d458bf731.

There is no safe way to remove this.

The problem appears when the first line of the script (or around the
first line) there is a line that sets the key input state. This line may
be executed before or after the input is initialized in the main loop,
and it isn't trivial to fix it.

Instead of fixing it, it's better to ask the scripts to wait for at
least one frame before setting the key input state.